# api/openapi.yml
openapi: 3.0.3
info:
  title: Fansworld Creator Platform API
  version: 1.0.0
  description: |
    Production-ready API for creator subscription platform with OnlyFans-level functionality.
    Supports content creation, subscription management, messaging, and creator monetization.
  contact:
    name: Fansworld API Team
    email: api@fansworld.com
  license:
    name: Proprietary
    url: https://fansworld.com/license
servers:
  - url: https://api.fansworld.com/v1
    description: Production server
  - url: https://staging-api.fansworld.com/v1
    description: Staging server
  - url: http://localhost:3001/api/v1
    description: Development server
security:
  - bearerAuth: []
paths:
...
  /auth/register:
    post:
      summary: Register new user account
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, username]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
                username: { type: string, minLength: 3, maxLength: 30 }
                referralCode: { type: string }
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Username or email already exists
  /auth/login:
    post:
      summary: Authenticate user
      tags: [Authentication]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Invalid credentials

  # User Management
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    AuthResponse:
      type: object
      properties:
        user: { $ref: '#/components/schemas/User' }
        accessToken: { type: string }
        refreshToken: { type: string }
        expiresIn: { type: integer }
    User:
      type: object
      properties:
        id: { type: string }
        email: { type: string, format: email }
        username: { type: string }
        displayName: { type: string }
        avatarUrl: { type: string, format: uri }
        role: { type: string, enum: [fan, creator, moderator, admin] }
        verified: { type: boolean }
        createdAt: { type: string, format: date-time }
...
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object }
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...